requirements.txt
binance-futures-connector==4.1.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
websocket-client==1.6.4 